Move scripts to the top-level
